% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixGGM.R, R/MixMCMC.R, R/MixMPLN.R
\name{compute_invSigma_mix}
\alias{compute_invSigma_mix}
\title{A function to}
\usage{
compute_invSigma_mix(wij, lambda, K)

compute_invSigma_mix(wij, lambda, K)

compute_invSigma_mix(wij, lambda, K)
}
\value{
invSigma

invSigma

invSigma
}
\description{
A function to

A function to

A function to
}
