% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R, R/generate.R
\name{generate_random_pd_sym_matrix}
\alias{generate_random_pd_sym_matrix}
\title{A function to}
\usage{
generate_random_pd_sym_matrix(nrow, sparse, min, max)

generate_random_pd_sym_matrix(nrow, sparse, min, max)
}
\arguments{
\item{nrow}{}

\item{min}{}

\item{nrow}{}

\item{min}{}
}
\value{
M

M
}
\description{
A function to

A function to
}
