% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixGGM.R, R/MixMCMC.R, R/MixMPLN.R
\name{huge_estimation}
\alias{huge_estimation}
\title{A function to}
\usage{
huge_estimation(lam, X1)

huge_estimation(lam, X1)

huge_estimation(lam, X1)
}
\arguments{
\item{lam}{}

\item{X1}{}

\item{lam}{}
}
\value{
as.matrix(inv_sigma_huge)

as.matrix(inv_sigma_huge)

as.matrix(inv_sigma_huge)
}
\description{
A function to

A function to

A function to
}
