% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixGGM.R, R/MixMCMC.R, R/MixMPLN.R
\name{myfr}
\alias{myfr}
\title{A function to}
\usage{
myfr(CS, NS)

myfr(CS, NS)

myfr(CS, NS)
}
\arguments{
\item{CS}{}

\item{CS}{}

\item{CS}{}
}
\value{
val

val

val
}
\description{
A function to

A function to

A function to
}
