% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R, R/generate.R
\name{sampling}
\alias{sampling}
\title{A function to}
\usage{
sampling(y, MinValue, MaxValue)

sampling(y, MinValue, MaxValue)
}
\arguments{
\item{y}{}

\item{MaxValue}{}

\item{y}{}

\item{MaxValue}{}
}
\value{
x

x
}
\description{
A function to

A function to
}
